Split simulator into modules
